update lexer error message to use size_t types

+++ src/include/core.h
@@ -2,6 +2,7 @@
 #define CORE_H
 #include "gc.h"
 #include "token.h"
+#include <stddef.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -40,7 +41,7 @@ void FreeCore(PanktiCore *core);
 void RunCore(PanktiCore *core);
 // Show error
 void CoreError(PanktiCore *core, Token *token, const char *msg);
-void CoreLexerError(PanktiCore *core, long line, long col, const char *msg);
+void CoreLexerError(PanktiCore *core, size_t line, size_t col, const char *msg);
 #ifdef __cplusplus
 }
 #endif
